# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CyberArcade CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: getContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: trialAttended
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
        - name: minLeadScore
          in: query
          schema:
            type: string
        - name: maxLeadScore
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
...
  /contacts/bulk:
    patch:
      operationId: bulkUpdateContacts
      tags: [contacts]
      summary: Bulk update contacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateContactsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUpdateResult"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactInput"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /contacts/{id}/activities:
    get:
      operationId: getContactActivities
      tags: [contacts]
      summary: Get contact activities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createContactActivity
      tags: [contacts]
      summary: Log a contact activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /contacts/{id}/campaigns:
    get:
      operationId: getContactCampaigns
      tags: [contacts]
      summary: Get campaigns a contact is part of
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignLog"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    BulkUpdateResult:
      type: object
      properties:
        updated:
          type: integer
      required:
        - updated
    Contact:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        source:
          type: string
          enum: [trial_attendee, website, instagram, facebook, referral, walk_in, sms, manual]
        status:
          type: string
          enum: [cold_lead, warm_lead, trial_attended, interested, enrolled, churned]
        childName:
          type: ["string", "null"]
        childAge:
          type: ["integer", "null"]
        childGrade:
          type: ["string", "null"]
        trialDate:
          type: ["string", "null"]
        trialAttended:
          type: boolean
        trialBatch:
          type: ["string", "null"]
        enrolledDate:
          type: ["string", "null"]
        enrolledCourse:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        leadScore:
          type: integer
        lastContactedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - email
        - source
        - status
        - trialAttended
        - tags
        - leadScore
        - createdAt
        - updatedAt
    ContactListResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - contacts
        - total
        - page
        - totalPages
    CreateContactInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        source:
          type: string
          enum: [trial_attendee, website, instagram, facebook, referral, walk_in, sms, manual]
        status:
          type: string
          enum: [cold_lead, warm_lead, trial_attended, interested, enrolled, churned]
        childName:
          type: ["string", "null"]
        childAge:
          type: ["integer", "null"]
        childGrade:
          type: ["string", "null"]
        trialDate:
          type: ["string", "null"]
        trialAttended:
          type: boolean
        trialBatch:
          type: ["string", "null"]
        enrolledDate:
          type: ["string", "null"]
        enrolledCourse:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        leadScore:
          type: integer
        notes:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
      required:
        - firstName
        - lastName
        - email
        - source
        - status
    UpdateContactInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        source:
          type: string
          enum: [trial_attendee, website, instagram, facebook, referral, walk_in, sms, manual]
        status:
          type: string
          enum: [cold_lead, warm_lead, trial_attended, interested, enrolled, churned]
        childName:
          type: ["string", "null"]
        childAge:
          type: ["integer", "null"]
        childGrade:
          type: ["string", "null"]
        trialDate:
          type: ["string", "null"]
        trialAttended:
          type: boolean
        trialBatch:
          type: ["string", "null"]
        enrolledDate:
          type: ["string", "null"]
        enrolledCourse:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        leadScore:
          type: integer
        notes:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        lastContactedAt:
          type: ["string", "null"]
    BulkUpdateContactsInput:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
        status:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        campaignId:
          type: ["string", "null"]
      required:
        - ids
    Activity:
      type: object
      properties:
        id:
          type: string
        contactId:
          type: string
        type:
          type: string
          enum: [note, call, email_sent, whatsapp_sent, status_change, enrolled]
        description:
          type: string
        createdAt:
          type: string
        createdBy:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
      required:
        - id
        - contactId
        - type
        - description
        - createdAt
    CreateActivityInput:
      type: object
      properties:
        type:
          type: string
          enum: [note, call, email_sent, whatsapp_sent, status_change, enrolled]
        description:
          type: string
        createdBy:
          type: ["string", "null"]
      required:
        - type
        - description
...
    CampaignLog:
      type: object
      properties:
        id:
          type: string
        campaignId:
          type: string
        contactId:
          type: string
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        channel:
          type: string
        status:
          type: string
          enum: [sent, delivered, opened, clicked, replied, converted, failed]
        sentAt:
          type: ["string", "null"]
        openedAt:
          type: ["string", "null"]
        clickedAt:
          type: ["string", "null"]
        convertedAt:
          type: ["string", "null"]
      required:
        - id
        - campaignId
        - contactId
        - channel
        - status
...