# lib/api-zod/src/generated/types/createActivityInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberArcade CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActivityInputType } from "./createActivityInputType";

export interface CreateActivityInput {
  type: CreateActivityInputType;
  description: string;
  /** @nullable */
  createdBy?: string | null;
}