# lib/api-zod/src/generated/types/updateContactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberArcade CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateContactInputSource } from "./updateContactInputSource";
import type { UpdateContactInputStatus } from "./updateContactInputStatus";

export interface UpdateContactInput {
  firstName?: string;
  lastName?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  city?: string | null;
  source?: UpdateContactInputSource;
  status?: UpdateContactInputStatus;
  /** @nullable */
  childName?: string | null;
  /** @nullable */
  childAge?: number | null;
  /** @nullable */
  childGrade?: string | null;
  /** @nullable */
  trialDate?: string | null;
  trialAttended?: boolean;
  /** @nullable */
  trialBatch?: string | null;
  /** @nullable */
  enrolledDate?: string | null;
  /** @nullable */
  enrolledCourse?: string | null;
  tags?: string[];
  leadScore?: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  lastContactedAt?: string | null;
}